initial some resources added
