+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.JobMarket">
         <activity
-            android:name=".presentation.adddetailscreen.AddScreenActivity"
+            android:name=".adddetailscreen.addpresentation.AddScreenActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
