+++ app/src/main/java/com/maad/jobmarket/adddetailscreen/addpresentation/AddScreenActivity.kt
@@ -1,10 +1,7 @@
-package com.maad.jobmarket.presentation.adddetailscreen
+package com.maad.jobmarket.adddetailscreen.addpresentation
 
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import com.maad.jobmarket.R
 import dagger.hilt.android.AndroidEntryPoint
