+++ app/src/main/java/com/maad/jobmarket/adddetailscreen/addpresentation/addressdetail/UserAddressFragment.kt
@@ -1,4 +1,4 @@
-package com.maad.jobmarket.presentation.adddetailscreen.userdetail
+package com.maad.jobmarket.adddetailscreen.addpresentation.addressdetail
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,15 +7,15 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.maad.jobmarket.R
 
-class UserDetailFragment : Fragment() {
-
+class UserAddressFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_user_detail, container, false)
+        return inflater.inflate(R.layout.fragment_user_address, container, false)
     }
+
 
 }
