+++ app/src/main/java/com/maad/jobmarket/adddetailscreen/addpresentation/resumedetail/UserResumeFragment.kt
@@ -1,4 +1,4 @@
-package com.maad.jobmarket.presentation.adddetailscreen.addressdetail
+package com.maad.jobmarket.adddetailscreen.addpresentation.resumedetail
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,14 +7,14 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.maad.jobmarket.R
 
-class UserAddressFragment : Fragment() {
+class UserResumeFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_user_address, container, false)
+        return inflater.inflate(R.layout.fragment_user_resume, container, false)
     }
