+++ app/src/main/java/com/maad/jobmarket/adddetailscreen/addpresentation/userdetail/UserDetailFragment.kt
@@ -1,4 +1,4 @@
-package com.maad.jobmarket.presentation.adddetailscreen.resumedetail
+package com.maad.jobmarket.adddetailscreen.addpresentation.userdetail
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,15 +7,15 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.maad.jobmarket.R
 
-class UserResumeFragment : Fragment() {
+class UserDetailFragment : Fragment() {
+
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_user_resume, container, false)
+        return inflater.inflate(R.layout.fragment_user_detail, container, false)
     }
-
 
 }
