+++ app/src/main/res/color/power_spinner_selector.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <!-- pressed state -->
+    <item android:color="@color/power_spinner_border_active" android:state_enabled="true" android:state_window_focused="true" android:state_pressed="true" />
+    <!-- unselected state -->
+    <item android:color="@color/power_spinner_border_in_active" android:state_enabled="true" android:state_window_focused="true" />
+    <!-- default -->
+    <item android:color="@color/power_spinner_border_active"/>
+</selector>
