+++ app/src/main/res/layout/activity_add_screen.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.adddetailscreen.AddScreenActivity">
+    tools:context=".adddetailscreen.addpresentation.AddScreenActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragmentContainerView"
