+++ app/src/main/res/layout/fragment_user_academic.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/background_color"
-    tools:context=".presentation.adddetailscreen.academicdetail.UserAcademicFragment">
+    tools:context=".adddetailscreen.addpresentation.academicdetail.UserAcademicFragment">
 
     <ImageView
         android:id="@+id/ivPopOut"
