+++ app/src/main/res/layout/fragment_user_address.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/background_color"
-    tools:context=".presentation.adddetailscreen.addressdetail.UserAddressFragment">
+    tools:context=".adddetailscreen.addpresentation.addressdetail.UserAddressFragment">
 
     <ImageView
         android:id="@+id/ivPopOut"
