+++ app/src/main/res/layout/fragment_user_detail.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/background_color"
-    tools:context=".presentation.adddetailscreen.userdetail.UserDetailFragment">
+    tools:context=".adddetailscreen.addpresentation.userdetail.UserDetailFragment">
 
     <TextView
         android:id="@+id/tvDetailHeading"
